rm DB interface from privacyenabledstate pkg (#1242)
rm privacyenabledstate.DB interface

As part of FAB-10794, we would be making changes to
LoadCommittedVersions(). Before doing that, we remove unnecessary
interface from the privacyenabledstate which includes
LoadCommittedVersions() API.

After we have removed privacyenabledstate.DB interface, the filename
db.go does not make sense. Hence, we rename it to types.go as it holds
all private data related types. Further, db_test.go contains tests that
should be in common_storage_db_test.go. Hence, we also move those test
to the correct location and rename db_test.go to types_test.go.

Further, we have renamed CommonStorageDB to DB, removed
DBProvider interface, and removed the term CommonStorage from variables.

FAB-17880


rename common_storage_db.go to db.go
