+++ core/ledger/kvledger/txmgmt/privacyenabledstate/db_test.go
@@ -11,7 +11,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
-	"os"
 	"testing"
 
 	"github.com/hyperledger/fabric-protos-go/peer"
@@ -20,58 +19,70 @@ import (
 	"github.com/hyperledger/fabric/core/ledger/cceventmgmt"
 	"github.com/hyperledger/fabric/core/ledger/internal/version"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/statedb"
+	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/statedb/statecouchdb"
+	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/statedb/stateleveldb"
+	"github.com/hyperledger/fabric/core/ledger/mock"
 	"github.com/hyperledger/fabric/core/ledger/util"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func TestMain(m *testing.M) {
-	exitCode := m.Run()
-	for _, testEnv := range testEnvs {
-		testEnv.StopExternalResource()
+func TestHealthCheckRegister(t *testing.T) {
+	fakeHealthCheckRegistry := &mock.HealthCheckRegistry{}
+	dbProvider := &DBProvider{
+		VersionedDBProvider: &stateleveldb.VersionedDBProvider{},
+		HealthCheckRegistry: fakeHealthCheckRegistry,
 	}
-	os.Exit(exitCode)
+
+	err := dbProvider.RegisterHealthChecker()
+	require.NoError(t, err)
+	require.Equal(t, 0, fakeHealthCheckRegistry.RegisterCheckerCallCount())
+
+	dbProvider.VersionedDBProvider = &statecouchdb.VersionedDBProvider{}
+	err = dbProvider.RegisterHealthChecker()
+	require.NoError(t, err)
+	require.Equal(t, 1, fakeHealthCheckRegistry.RegisterCheckerCallCount())
+
+	arg1, arg2 := fakeHealthCheckRegistry.RegisterCheckerArgsForCall(0)
+	require.Equal(t, "couchdb", arg1)
+	require.NotNil(t, arg2)
 }
 
-func TestBatch(t *testing.T) {
-	batch := UpdateMap(make(map[string]NsBatch))
-	v := version.NewHeight(1, 1)
-	for i := 0; i < 5; i++ {
-		for j := 0; j < 5; j++ {
-			for k := 0; k < 5; k++ {
-				batch.Put(fmt.Sprintf("ns-%d", i), fmt.Sprintf("collection-%d", j), fmt.Sprintf("key-%d", k),
-					[]byte(fmt.Sprintf("value-%d-%d-%d", i, j, k)), v)
-			}
-		}
+func TestGetIndexInfo(t *testing.T) {
+	chaincodeIndexPath := "META-INF/statedb/couchdb/indexes/indexColorSortName.json"
+	actualIndexInfo := getIndexInfo(chaincodeIndexPath)
+	expectedIndexInfo := &indexInfo{
+		hasIndexForChaincode:  true,
+		hasIndexForCollection: false,
+		collectionName:        "",
 	}
-	for i := 0; i < 5; i++ {
-		for j := 0; j < 5; j++ {
-			for k := 0; k < 5; k++ {
-				vv := batch.Get(fmt.Sprintf("ns-%d", i), fmt.Sprintf("collection-%d", j), fmt.Sprintf("key-%d", k))
-				assert.NotNil(t, vv)
-				assert.Equal(t,
-					&statedb.VersionedValue{Value: []byte(fmt.Sprintf("value-%d-%d-%d", i, j, k)), Version: v},
-					vv)
-			}
-		}
+	require.Equal(t, expectedIndexInfo, actualIndexInfo)
+
+	collectionIndexPath := "META-INF/statedb/couchdb/collections/collectionMarbles/indexes/indexCollMarbles.json"
+	actualIndexInfo = getIndexInfo(collectionIndexPath)
+	expectedIndexInfo = &indexInfo{
+		hasIndexForChaincode:  false,
+		hasIndexForCollection: true,
+		collectionName:        "collectionMarbles",
 	}
-	assert.Nil(t, batch.Get("ns-1", "collection-1", "key-5"))
-	assert.Nil(t, batch.Get("ns-1", "collection-5", "key-1"))
-	assert.Nil(t, batch.Get("ns-5", "collection-1", "key-1"))
-}
+	require.Equal(t, expectedIndexInfo, actualIndexInfo)
 
-func TestHashBatchContains(t *testing.T) {
-	batch := NewHashedUpdateBatch()
-	batch.Put("ns1", "coll1", []byte("key1"), []byte("val1"), version.NewHeight(1, 1))
-	assert.True(t, batch.Contains("ns1", "coll1", []byte("key1")))
-	assert.False(t, batch.Contains("ns1", "coll1", []byte("key2")))
-	assert.False(t, batch.Contains("ns1", "coll2", []byte("key1")))
-	assert.False(t, batch.Contains("ns2", "coll1", []byte("key1")))
+	incorrectChaincodeIndexPath := "META-INF/statedb/couchdb/indexColorSortName.json"
+	actualIndexInfo = getIndexInfo(incorrectChaincodeIndexPath)
+	expectedIndexInfo = &indexInfo{
+		hasIndexForChaincode:  false,
+		hasIndexForCollection: false,
+		collectionName:        "",
+	}
+	require.Equal(t, expectedIndexInfo, actualIndexInfo)
 
-	batch.Delete("ns1", "coll1", []byte("deleteKey"), version.NewHeight(1, 1))
-	assert.True(t, batch.Contains("ns1", "coll1", []byte("deleteKey")))
-	assert.False(t, batch.Contains("ns1", "coll1", []byte("deleteKey1")))
-	assert.False(t, batch.Contains("ns1", "coll2", []byte("deleteKey")))
-	assert.False(t, batch.Contains("ns2", "coll1", []byte("deleteKey")))
+	incorrectCollectionIndexPath := "META-INF/statedb/couchdb/collections/indexes/indexCollMarbles.json"
+	actualIndexInfo = getIndexInfo(incorrectCollectionIndexPath)
+	require.Equal(t, expectedIndexInfo, actualIndexInfo)
+
+	incorrectIndexPath := "META-INF/statedb/"
+	actualIndexInfo = getIndexInfo(incorrectIndexPath)
+	require.Equal(t, expectedIndexInfo, actualIndexInfo)
 }
 
 func TestDB(t *testing.T) {
@@ -97,8 +108,7 @@ func testDB(t *testing.T, env TestEnv) {
 	putPvtUpdates(t, updates, "ns1", "coll1", "key2", []byte("pvt_value2"), version.NewHeight(1, 5))
 	putPvtUpdates(t, updates, "ns2", "coll1", "key3", []byte("pvt_value3"), version.NewHeight(1, 6))
 	db.ApplyPrivacyAwareUpdates(updates, version.NewHeight(2, 6))
-	commonStorageDB := db.(*CommonStorageDB)
-	bulkOptimizable, ok := commonStorageDB.VersionedDB.(statedb.BulkOptimizable)
+	bulkOptimizable, ok := db.VersionedDB.(statedb.BulkOptimizable)
 	if ok {
 		bulkOptimizable.ClearCachedVersions()
 	}
@@ -242,7 +252,7 @@ func testGetStateRangeScanIterator(t *testing.T, env TestEnv) {
 
 func TestQueryOnCouchDB(t *testing.T) {
 	for _, env := range testEnvs {
-		_, ok := env.(*CouchDBCommonStorageTestEnv)
+		_, ok := env.(*CouchDBTestEnv)
 		if !ok {
 			continue
 		}
@@ -323,7 +333,7 @@ func testQueryOnCouchDB(t *testing.T, env TestEnv) {
 
 func TestLongDBNameOnCouchDB(t *testing.T) {
 	for _, env := range testEnvs {
-		_, ok := env.(*CouchDBCommonStorageTestEnv)
+		_, ok := env.(*CouchDBTestEnv)
 		if !ok {
 			continue
 		}
@@ -396,31 +406,9 @@ func testKey(i int) string {
 	return fmt.Sprintf("key%d", i)
 }
 
-func TestCompositeKeyMap(t *testing.T) {
-	b := NewPvtUpdateBatch()
-	b.Put("ns1", "coll1", "key1", []byte("testVal1"), nil)
-	b.Delete("ns1", "coll2", "key2", nil)
-	b.Put("ns2", "coll1", "key1", []byte("testVal3"), nil)
-	b.Put("ns2", "coll2", "key2", []byte("testVal4"), nil)
-	m := b.ToCompositeKeyMap()
-	assert.Len(t, m, 4)
-	vv, ok := m[PvtdataCompositeKey{"ns1", "coll1", "key1"}]
-	assert.True(t, ok)
-	assert.Equal(t, []byte("testVal1"), vv.Value)
-	vv, ok = m[PvtdataCompositeKey{"ns1", "coll2", "key2"}]
-	assert.Nil(t, vv.Value)
-	assert.True(t, ok)
-	_, ok = m[PvtdataCompositeKey{"ns2", "coll1", "key1"}]
-	assert.True(t, ok)
-	_, ok = m[PvtdataCompositeKey{"ns2", "coll2", "key2"}]
-	assert.True(t, ok)
-	_, ok = m[PvtdataCompositeKey{"ns2", "coll1", "key8888"}]
-	assert.False(t, ok)
-}
-
 func TestHandleChainCodeDeployOnCouchDB(t *testing.T) {
 	for _, env := range testEnvs {
-		_, ok := env.(*CouchDBCommonStorageTestEnv)
+		_, ok := env.(*CouchDBTestEnv)
 		if !ok {
 			continue
 		}
@@ -453,8 +441,6 @@ func testHandleChainCodeDeploy(t *testing.T, env TestEnv) {
 	ccp := &peer.CollectionConfigPackage{Config: []*peer.CollectionConfig{coll1}}
 	chaincodeDef := &cceventmgmt.ChaincodeDefinition{Name: "ns1", Hash: nil, Version: "", CollectionConfigs: ccp}
 
-	commonStorageDB := db.(*CommonStorageDB)
-
 	// Test indexes for side databases
 	dbArtifactsTarBytes := testutil.CreateTarBytesForTest(
 		[]*testutil.TarFileEntry{
@@ -485,7 +471,7 @@ func testHandleChainCodeDeploy(t *testing.T, env TestEnv) {
 
 	// The collection config is added to the chaincodeDef but missing collectionMarblesPrivateDetails.
 	// Hence, the index on collectionMarblesPrivateDetails cannot be created
-	err = commonStorageDB.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
+	err = db.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
 	assert.NoError(t, err)
 
 	coll2 := createCollectionConfig("collectionMarblesPrivateDetails")
@@ -496,7 +482,7 @@ func testHandleChainCodeDeploy(t *testing.T, env TestEnv) {
 	// including collectionMarblesPrivateDetails which was missing earlier.
 	// Hence, the existing indexes must be updated and the new index must be created for
 	// collectionMarblesPrivateDetails
-	err = commonStorageDB.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
+	err = db.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
 	assert.NoError(t, err)
 
 	chaincodeDef = &cceventmgmt.ChaincodeDefinition{Name: "ns1", Hash: nil, Version: "", CollectionConfigs: nil}
@@ -505,19 +491,19 @@ func testHandleChainCodeDeploy(t *testing.T, env TestEnv) {
 	// process reads the collection config from state db. However, the state db does not contain
 	// any collection config for this chaincode. Hence, index creation/update on all collections
 	// should fail
-	err = commonStorageDB.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
+	err = db.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
 	assert.NoError(t, err)
 
 	//Test HandleChaincodeDefinition with a nil tar file
-	err = commonStorageDB.HandleChaincodeDeploy(chaincodeDef, nil)
+	err = db.HandleChaincodeDeploy(chaincodeDef, nil)
 	assert.NoError(t, err)
 
 	//Test HandleChaincodeDefinition with a bad tar file
-	err = commonStorageDB.HandleChaincodeDeploy(chaincodeDef, []byte(`This is a really bad tar file`))
+	err = db.HandleChaincodeDeploy(chaincodeDef, []byte(`This is a really bad tar file`))
 	assert.NoError(t, err, "Error should not have been thrown for a bad tar file")
 
 	//Test HandleChaincodeDefinition with a nil chaincodeDef
-	err = commonStorageDB.HandleChaincodeDeploy(nil, dbArtifactsTarBytes)
+	err = db.HandleChaincodeDeploy(nil, dbArtifactsTarBytes)
 	assert.Error(t, err, "Error should have been thrown for a nil chaincodeDefinition")
 
 	// Create a tar file for test with 2 index definitions - one of them being errorneous
@@ -536,7 +522,7 @@ func testHandleChainCodeDeploy(t *testing.T, env TestEnv) {
 	// There should be 1 entry
 	assert.Len(t, fileEntries, 1)
 
-	err = commonStorageDB.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
+	err = db.HandleChaincodeDeploy(chaincodeDef, dbArtifactsTarBytes)
 	assert.NoError(t, err)
 
 }
